+++ src/TitleDetails/AddPieceModal/AddPieceModalContainer.test.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import { render, fireEvent } from '@testing-library/react';
+import { render, screen, fireEvent } from '@testing-library/react';
+import user from '@testing-library/user-event';
 import { IntlProvider } from 'react-intl';
 import { MemoryRouter } from 'react-router-dom';
 
@@ -13,51 +14,31 @@ jest.mock('@folio/stripes-acq-components', () => ({
 jest.mock('../../common/components/LineLocationsView/LineLocationsView',
   () => jest.fn().mockReturnValue('LineLocationsView'));
 
-const renderAddPieceModalContainer = (
-  close,
-  onSubmit,
-  initialValues,
-  instanceId,
-  onCheckIn,
-  poLine,
-  locations,
-  locationIds,
-) => (render(
+const defaultProps = {
+  close: jest.fn(),
+  onSubmit: jest.fn(),
+  onCheckIn: jest.fn(),
+  deletePiece: jest.fn(),
+  poLine: { id: 'poLineId', physical: { createInventory: 'None' }, locations: [{ locationId: '001' }] },
+  initialValues: { enumeration: 'testenumeration', format: 'Physical', id: 'id', poLineId: 'poLineId', titleId: 'titleId', locationId: '001' },
+  locations: [{ name: 'Location', code: 'code', id: '001' }],
+  locationIds: ['001'],
+};
+
+const renderAddPieceModalContainer = (props = {}) => render(
   <IntlProvider locale="en">
     <MemoryRouter>
       <AddPieceModalContainer
-        close={close}
-        initialValues={initialValues}
-        instanceId={instanceId}
-        locations={locations}
-        locationIds={locationIds}
-        onCheckIn={onCheckIn}
-        onSubmit={onSubmit}
-        poLine={poLine}
-        deletePiece={jest.fn}
+        {...defaultProps}
+        {...props}
       />
     </MemoryRouter>
   </IntlProvider>,
-));
+);
 
 describe('AddPieceModalContainer', () => {
-  let close;
-  let onSubmit;
-  let onCheckIn;
-
-  beforeEach(() => {
-    close = jest.fn();
-    onSubmit = jest.fn();
-    onCheckIn = jest.fn();
-  });
-
   it('should display Edit Piece form', () => {
-    const poLine = { id: 'poLineId', physical: { createInventory: 'None' }, locations: [{ locationId: '001' }] };
-    const initialValues = { enumeration: 'testenumeration', format: 'Physical', id: 'id', poLineId: 'poLineId', titleId: 'titleId', locationId: '001' };
-    const locations = [{ name: 'Location', code: 'code', id: '001' }];
-    const locationIds = ['001'];
-
-    const { getByLabelText, getByText, queryByText } = renderAddPieceModalContainer(close, onSubmit, initialValues, 'instanceId', onCheckIn, poLine, locations, locationIds);
+    const { getByLabelText, getByText, queryByText } = renderAddPieceModalContainer();
 
     // header is rendered
     expect(getByText('ui-receiving.piece.caption')).toBeDefined();
@@ -73,31 +54,13 @@ describe('AddPieceModalContainer', () => {
   });
 
   it('should display Edit Received Piece form', () => {
-    const poLine = { id: 'poLineId', physical: { createInventory: 'None' }, locations: [{ locationId: '001' }] };
-    const piece = {
-      caption: 'testcaption',
-      copyNumber: 'testcopynumber',
-      enumeration: 'testenumeration',
-      format: 'Physical',
-      id: 'id',
-      locationId: '001',
-      poLineId: 'poLineId',
-      titleId: 'titleId',
-      receivingStatus: PIECE_STATUS.received,
-    };
-    const locations = [{ name: 'Location', id: '001' }];
-    const locationIds = ['001'];
-
-    const { getByLabelText, queryByText, getByText } = renderAddPieceModalContainer(
-      close,
-      onSubmit,
-      piece,
-      'instanceId',
-      onCheckIn,
-      poLine,
-      locations,
-      locationIds,
-    );
+    const { getByLabelText, queryByText, getByText } = renderAddPieceModalContainer({
+      instanceId: 'instanceId',
+      initialValues: {
+        ...defaultProps.initialValues,
+        receivingStatus: PIECE_STATUS.received,
+      },
+    });
 
     expect(getByLabelText('ui-receiving.piece.caption').disabled).toBeFalsy();
     expect(getByLabelText('ui-receiving.piece.copyNumber').disabled).toBeFalsy();
@@ -105,5 +68,16 @@ describe('AddPieceModalContainer', () => {
     expect(getByText('stripes-acq-components.piece.pieceFormat.physical')).toBeDefined();
     expect(getByLabelText('ui-receiving.piece.receiptDate').disabled).toBeFalsy();
     expect(queryByText('ui-receiving.piece.actions.quickReceive')).toBeFalsy();
+  });
+
+  it('should call on submit when \'Save\' button was clicked', async () => {
+    renderAddPieceModalContainer();
+
+    const saveBtn = await screen.findByRole('button', {
+      name: 'ui-receiving.piece.actions.save',
+    });
+
+    user.click(saveBtn);
+    expect(defaultProps.onSubmit).toHaveBeenCalled();
   });
 });